requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# intex/server/main.py: 2
aiofiles == 0.7.0

# intex/server/main.py: 11,12,13,14
fastapi == 0.58.0

# intex/data/generate_review_metadata.py: 12,14
gensim == 4.0.1

# intex/evaluate_deep_agent.py: 8
# intex/intex_env/intex_env/__init__.py: 1
# intex/intex_env/intex_env/envs/intex_env.py: 9
# intex/produce_execution_data.py: 3
# intex/train_deep_agent.py: 19
gym == 0.17.3

# intex/utilities/save_and_smooth_and_plot.py: 3
matplotlib == 3.4.3

# intex/data/generate_review_metadata.py: 17
nltk == 3.6.2

# intex/data/generate_review_metadata.py: 6,13
# intex/evaluate_deep_agent.py: 12,17
# intex/intex_env/intex_env/envs/intex_env.py: 10
# intex/produce_execution_data.py: 4
# intex/rl/agent.py: 9
# intex/rl/boolean_state_features.py: 8
# intex/rl/state_features.py: 8
# intex/server/custom_json_encoder.py: 2
# intex/server/main.py: 8
# intex/train_deep_agent.py: 23
# intex/train_policy.py: 8
# intex/utilities.py: 4,5
numpy == 1.19.5

# intex/data/db_interface.py: 4
# intex/data/generate_review_metadata.py: 22
# intex/evaluate_deep_agent.py: 16
# intex/exploration_actions.py: 11
# intex/import_amazon_data.py: 2
# intex/import_imdb_data.py: 1,2
# intex/intex_env/intex_env/envs/intex_env.py: 11
# intex/rl/boolean_state_features.py: 3,9
# intex/rl/state_features.py: 3,9
# intex/utilities/save_and_smooth_and_plot.py: 1
# intex/utilities/smooth_csv.py: 2
pandas == 1.0.5

# intex/evaluate_deep_agent.py: 7
# intex/produce_execution_data.py: 5
# intex/rl/deep_q.py: 4
# intex/train_deep_agent.py: 28
pfrl == 0.3.0

# intex/data/db_interface.py: 3
# intex/generate_target_items.py: 3
psycopg2_binary == 2.8.6

# intex/import_imdb_data.py: 4
requests == 2.25.1

# intex/data/generate_review_metadata.py: 18
scipy == 1.7.0

# intex/intex_env/setup.py: 1
setuptools == 52.0.0

# intex/data/generate_review_metadata.py: 19,20
spacy == 3.0.6

# intex/server/main.py: 15
starlette == 0.13.4

# intex/data/generate_review_metadata.py: 21
textblob == 0.15.3

# intex/evaluate_deep_agent.py: 11
# intex/produce_execution_data.py: 6
# intex/rl/deep_q.py: 5,6
# intex/train_deep_agent.py: 16,17,18
torch == 1.8.1

# intex/data/generate_review_metadata.py: 23
# intex/import_amazon_data.py: 5
# intex/import_imdb_data.py: 3
# intex/rl/agent.py: 11
tqdm == 4.61.2

# intex/start-server.py: 1
uvicorn == 0.11.5

# intex/evaluate_deep_agent.py: 13
# intex/train_deep_agent.py: 25
# intex/utilities/save_and_smooth_and_plot.py: 2
wandb == 0.10.11
